requirements.txt
allure-pytest
requests~=2.26.0
pytest~=6.2.4
flaky~=3.7.0